Updated dstar code.
Used the Robotics Toolbox in Matlab to do the path planning using the D*
algorithm.
It takes an Occupancy Grid, the present location, and the destination
from ROS.
The algorithm outputs a vector with a sequency of points on the
occupancy grid to go from the present location to the destination.