Add files via upload
In this version, the event handler scans points in a small region near the location of the boat. The area is approximately the probability triangle given by the camera. The code scans for points along rays starting from the upper boundary of the triangle and de-incrementing the angle till it reaches the lower bound. Angles of objects are calculated by both camera and lidar. The camera angle will be subscribed directly, while the lidar angle will be calculated from the difference between the orientation of the boat and the orientation of the object with respect to (wrt) the x-axis. Then the camera and lidar object angle will be compared for errors.